pyproject.toml
[project]
name = "bicameral-manifold-gmoe"
version = "0.1.0"
description = "Official implementation of the Fractal Bottleneck Hypothesis and Dual-Manifold Architecture using BitNet b1.58."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-cpp-python>=0.3.0",
    "huggingface-hub",
    "rich",
    "numpy",
    "scipy",
    "click",
    "tqdm",
    "textual>=0.40.0",
    "plotext>=5.2.8",
    "textual-plotext>=0.2.1",
    "psutil>=7.2.1",
    "torch",
    "transformers",
    "safetensors",
    "sentencepiece",
    "datasets",
    "protobuf>=6.33.2",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "cli"]
